updated base client
